+++ DataFormats/HcalCalibObjects/BuildFile.xml
@@ -1,5 +1,6 @@
 <use   name="DataFormats/Common"/>
-
+<use   name="FWCore/ParameterSet"/>
+<use   name="FWCore/Framework"/>
 <export>
   <lib   name="1"/>
 </export>
