+++ DataFormats/HcalCalibObjects/src/classes_def.xml
@@ -2,9 +2,6 @@
   <class name="HEDarkening" ClassVersion="0">
    <version ClassVersion="0" checksum="3959036241"/>
   </class>
-  <class name="HFRecalibration" ClassVersion="0">
-   <version ClassVersion="0" checksum="526251233"/>
-  </class>
   <class name="HOCalibVariables" ClassVersion="16">
    <version ClassVersion="16" checksum="2161587371"/>
   </class>
